requirements.txt
Django>=4.2.0
graphene-django>=3.0.0
django-filter>=22.1
django-crontab==0.7.1
gql[requests]>=3.4.0
celery>=5.2.0
django-celery-beat>=2.4.0
redis>=4.0.0